+++ demos/react-native-supabase-background-sync/powersync/AppSchema.ts
@@ -1,0 +1,35 @@
+import { column, Schema, Table } from '@powersync/react-native';
+
+export const LIST_TABLE = 'lists';
+export const TODO_TABLE = 'todos';
+
+const lists = new Table({
+  created_at: column.text,
+  name: column.text,
+  owner_id: column.text,
+  user_id: column.text
+});
+
+const todos = new Table(
+  {
+    list_id: column.text,
+    created_at: column.text,
+    completed_at: column.text,
+    description: column.text,
+    created_by: column.text,
+    completed_by: column.text,
+    completed: column.integer,
+    user_id: column.text
+  },
+  { indexes: { list: ['list_id'] } }
+);
+
+export const AppSchema = new Schema({
+  todos,
+  lists,
+});
+
+// For types
+export type Database = (typeof AppSchema)['types'];
+export type TodoRecord = Database['todos'];
+export type ListRecord = Database['lists'];