+++ demos/react-native-supabase-todolist/library/powersync/AppSchema.ts
@@ -1,5 +1,5 @@
-import { AttachmentTable } from '@powersync/attachments';
 import { column, Schema, Table } from '@powersync/react-native';
+import { AttachmentTable } from '@powersync/common';
 
 export const LIST_TABLE = 'lists';
 export const TODO_TABLE = 'todos';
@@ -27,9 +27,7 @@ const lists = new Table({
 export const AppSchema = new Schema({
   todos,
   lists,
-  attachments: new AttachmentTable({
-    name: 'attachments',
-  }),
+  attachments: new AttachmentTable(),
 });
 
 export type Database = (typeof AppSchema)['types'];