+++ packages/web/vitest.config.ts
@@ -1,6 +1,4 @@
 import path from 'path';
-import topLevelAwait from 'vite-plugin-top-level-await';
-import wasm from 'vite-plugin-wasm';
 import { defineConfig, UserConfigExport } from 'vitest/config';
 
 const config: UserConfigExport = {
@@ -19,23 +17,16 @@ const config: UserConfigExport = {
        * they link to `.js` files.
        */
       '@powersync/web': path.resolve(__dirname, './lib/src'),
-      // https://jira.mongodb.org/browse/NODE-5773
-      bson: import.meta.resolve('bson'),
       // Mock WebRemote to throw 401 errors for all HTTP requests in tests
       '../../db/sync/WebRemote.js': path.resolve(__dirname, './tests/mocks/MockWebRemote.ts')
     }
   },
   worker: {
-    format: 'es',
-    plugins: () => [wasm(), topLevelAwait()]
+    format: 'es'
   },
   optimizeDeps: {
-    // Don't optimise these packages as they contain web workers and WASM files.
-    // https://github.com/vitejs/vite/issues/11672#issuecomment-1415820673
-    exclude: ['@journeyapps/wa-sqlite', '@powersync/web'],
-    include: []
+    exclude: ['@journeyapps/wa-sqlite']
   },
-  plugins: [wasm(), topLevelAwait()],
   test: {
     globals: true,
     include: ['tests/**/*.test.ts'],